Update _garage-auto-close.js
Modified Garage Auto Close to use nodeAtScheduleTime instead of retrieving the node from the database.  If the node is modified by the event, return the modified node.  The property modifiedByEvent will be set to true, this will indicate to the schedule event the node has been modified and should be updated in the database.